pyproject.toml
[project]
name = "rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chromadb>=1.5.7",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.28",
    "langchain-text-splitters>=1.1.1",
    "langgraph>=1.1.6",
    "langsmith>=0.7.29",
    "lxml>=6.0.2",
    "marker-pdf>=1.10.2",
    "onnx>=1.21.0",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=10.4.0",
    "sentence-transformers>=5.3.0",
    "tavily-python>=0.7.23",
]

[dependency-groups]
dev = [
    "black>=26.3.1",
    "isort>=8.0.1",
    "rough>=1.6",
]
